requirements.txt
moviepy>=1.0.3
pillow>=10.0.0
imageio>=2.34.0
imageio-ffmpeg>=0.4.9
numpy>=1.25.0
pyinstaller>=6.0.0

